pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ani-sage"
version = "0.1.0"
authors = [
  { name="tendev", email="tendev@example.com" },
]
description = "An AI-powered anime recommendation system"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.6.1",
    "aiohttp>=3.9.3",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "strict"
addopts = "-v"
markers = [
    "integration: marks tests as integration tests"
]

[project.urls]
"Homepage" = "https://github.com/tendev/ani-sage"
"Bug Tracker" = "https://github.com/tendev/ani-sage/issues"
